Add THTensorMathSIMD.c
