+++ lib/TH/CMakeLists.txt
@@ -377,6 +377,7 @@ INSTALL(FILES
   generic/THTensorLapack.c
   generic/THTensorLapack.h
   generic/THTensorMath.c
+  generic/THTensorMathSIMD.c
   generic/THTensorMathDispatch.c
   generic/THTensorMath.h
   generic/THTensorRandom.c
