+++ lib/TH/THTensor.c
@@ -27,6 +27,9 @@
 #include "generic/THTensorMath.c"
 #include "THGenerateAllTypes.h"
 
+#include "generic/THTensorMathSIMD.c"
+#include "THGenerateAllTypes.h"
+
 #include "generic/THTensorMathDispatch.c"
 #include "THGenerateAllTypes.h"
